feat: add a copy button
